+++ GamblersMod/GrabbableObjectCustom/GrabbableObjectCustom.cs
@@ -1,0 +1,6 @@
+﻿namespace GamblersMod.GrabbableObjectCustom
+{
+    internal class GrabbableObjectCustom
+    {
+    }
+}