+++ GamblersMod/Player/PlayerGamblingUIManager.cs
@@ -44,7 +44,7 @@ namespace GamblersMod.Patches
             gamblingMachineInteractionText.text = interactionText;
             gamblingMachineInteractionText.alignment = TextAnchor.MiddleCenter;
             gamblingMachineInteractionText.font = Plugin.GamblingFont;
-            gamblingMachineInteractionText.rectTransform.sizeDelta = new Vector2(300, 200);
+            gamblingMachineInteractionText.rectTransform.sizeDelta = new Vector2(500, 400);
             gamblingMachineInteractionText.fontSize = 26;
 
             gamblingMachineInteractionText.transform.parent = gamblingMachineInteractionTextCanvasObject.transform;
@@ -59,8 +59,8 @@ namespace GamblersMod.Patches
             gamblingMachineInteractionScrapInfoText.text = interactionText;
             gamblingMachineInteractionScrapInfoText.alignment = TextAnchor.MiddleCenter;
             gamblingMachineInteractionScrapInfoText.font = Plugin.GamblingFont;
-            gamblingMachineInteractionScrapInfoText.rectTransform.sizeDelta = new Vector2(300, 200);
-            gamblingMachineInteractionScrapInfoText.fontSize = 16;
+            gamblingMachineInteractionScrapInfoText.rectTransform.sizeDelta = new Vector2(500, 300);
+            gamblingMachineInteractionScrapInfoText.fontSize = 18;
             gamblingMachineInteractionScrapInfoText.color = Color.green;
 
             gamblingMachineInteractionScrapInfoText.transform.parent = gamblingMachineInteractionTextCanvasObject.transform;